requirements.txt
click==7.1.1
gunicorn==19.9.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.0.0
Werkzeug==1.0.1
pytesseract==0.3.3
opencv-contrib-python-headless==4.2.0.32
opencv-python-headless==4.2.0.32
numpy==1.16.5
tabula-py== 2.1.0 
PyPDF2==1.26.0
pandas==1.0.3
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
docx2python==1.21
docxpy==0.8.5
imutils==0.5.3
glob2==0.7
python-docx==0.8.10
xlrd==1.1.0